chore: update tests and reorganize due to form changes for the exif data
